requirements.txt
aiofiles==24.1.0
aiogram==3.17.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aioredis==2.0.1
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
attrs==24.3.0
certifi==2024.12.14
click==8.1.8
exceptiongroup==1.2.2
fastapi==0.115.7
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
magic-filter==1.0.12
multidict==6.1.0
packaging==24.2
pluggy==1.5.0
propcache==0.2.1
pydantic==2.10.5
pydantic_core==2.27.2
pytest==8.3.4
python-dotenv==1.0.1
redis==5.2.1
setuptools==75.8.0
sniffio==1.3.1
starlette==0.45.2
typing_extensions==4.12.2
uvicorn==0.34.0
yarl==1.18.3
